Move code generation into the compiler.
